// src/api/generated/model/createIngredientBodyBaseUnit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */

export type CreateIngredientBodyBaseUnit = typeof CreateIngredientBodyBaseUnit[keyof typeof CreateIngredientBodyBaseUnit];


export const CreateIngredientBodyBaseUnit = {
  g: 'g',
  ml: 'ml',
} as const;